Refactored SecurityRules
